+++ R/reexports.R
@@ -1,0 +1,55 @@
+#' @export
+transpose <- purrr::transpose
+
+#' @importFrom dplyr first last mutate
+#' @export
+dplyr::first
+
+#' @export
+dplyr::last
+
+#' @export
+dplyr::mutate
+
+#' @export
+magrittr::`%>%`
+
+#' @export
+magrittr::`%<>%`
+
+#' @export
+magrittr::set_names
+
+#' @export
+purrr::map
+
+#' @importFrom data.table data.table is.data.table as.data.table
+#' @export
+data.table::data.table
+
+#' @export
+data.table::is.data.table
+
+#' @export
+data.table::as.data.table
+
+#' @importFrom lubridate make_date
+#' @export
+lubridate::make_date
+
+#' @importFrom foreach foreach %do% %dopar%
+#' @export 
+foreach::`%dopar%`
+
+#' @export 
+foreach::`%do%`
+
+#' @import iterators
+#' @export 
+iterators::iter
+
+#' @export
+iterators::icount
+
+# #' @export
+# lubridate::date