+++ NEWS.md
@@ -1,3 +1,11 @@
+
+# Ipaper 0.1.7
+
+- Many functions are reexported in Ipaper, including `iterators`, `foreach`, `lubridate`,
+  `data.table`, `dplyr`, `remotes`. One Ipaper is enough.
+- Addin `select_word` works
+- Git functions works, add `git_commit`, `git_commit_amend`, `git_push`
+
 # Ipaper 0.1.4.9000
 
 * Added a `NEWS.md` file to track changes to the package.
@@ -5,5 +13,5 @@
 
 # Ipaper 0.1.5.9000
 
-* `mkTRend_rcpp` is at least 6-fold faster
+* `mkTRend_rcpp` is at least 6-fold faster (moved to rtrend)
 * add `acf.fft`